[Github][Integration] Clean Up
